+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug fixes
 * Add seekable() method in S3Reader to eliminate tensor copies during DCP loading (#359)
+* Override S3Writer closed property and block writes after close (#360)
 
 ## v1.4.3 (July 25, 2025)